✨ feat: create mock for table
