+++ src/app/app.component.html
@@ -15,10 +15,10 @@
   <body>
     <ce-details-card
       [title]="'João Vitor Pereira dos Santos'"
-      [month]="2019 / 06 / 01"
+      [month]="date"
       [contentNumber]="50"
       [detailColor]="'primary'"
     ></ce-details-card>
-    <ce-table [ELEMENT_DATA]="ELEMENT_DATA"></ce-table>
+    <ce-table></ce-table>
   </body>
 </html>
